Reworked login with firebase api
