--- lib/appIcons_icons.dart
@@ -1,349 +1,0 @@
-/// Flutter icons Icons
-/// Copyright (C) 2020 by original authors @ fluttericon.com, fontello.com
-/// This font was generated by FlutterIcon.com, which is derived from Fontello.
-///
-/// To use this font, place it in your fonts/ directory and include the
-/// following in your pubspec.yaml
-///
-/// flutter:
-///   fonts:
-///    - family:  Icons
-///      fonts:
-///       - asset: fonts/Icons.ttf
-///
-/// 
-///
-import 'package:flutter/widgets.dart';
-
-class Icons {
-  Icons._();
-
-  static const _kFontFam = 'Icons';
-  static const _kFontPkg = null;
-
-  static const IconData uniF100 = IconData(0xf100, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF101 = IconData(0xf101, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF102 = IconData(0xf102, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF103 = IconData(0xf103, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF104 = IconData(0xf104, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF105 = IconData(0xf105, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF106 = IconData(0xf106, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF107 = IconData(0xf107, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF108 = IconData(0xf108, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF109 = IconData(0xf109, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF10A = IconData(0xf10a, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF10B = IconData(0xf10b, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF10C = IconData(0xf10c, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF10D = IconData(0xf10d, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF10E = IconData(0xf10e, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF10F = IconData(0xf10f, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF110 = IconData(0xf110, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF111 = IconData(0xf111, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF112 = IconData(0xf112, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF113 = IconData(0xf113, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF114 = IconData(0xf114, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF115 = IconData(0xf115, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF116 = IconData(0xf116, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF117 = IconData(0xf117, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF118 = IconData(0xf118, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF119 = IconData(0xf119, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF11A = IconData(0xf11a, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF11B = IconData(0xf11b, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF11C = IconData(0xf11c, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF11D = IconData(0xf11d, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF11E = IconData(0xf11e, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF11F = IconData(0xf11f, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF120 = IconData(0xf120, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF121 = IconData(0xf121, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF122 = IconData(0xf122, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF123 = IconData(0xf123, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF124 = IconData(0xf124, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF125 = IconData(0xf125, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF126 = IconData(0xf126, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF127 = IconData(0xf127, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF128 = IconData(0xf128, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF129 = IconData(0xf129, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF12A = IconData(0xf12a, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF12B = IconData(0xf12b, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF12C = IconData(0xf12c, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF12D = IconData(0xf12d, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF12E = IconData(0xf12e, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF12F = IconData(0xf12f, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF130 = IconData(0xf130, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF131 = IconData(0xf131, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF132 = IconData(0xf132, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF133 = IconData(0xf133, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF134 = IconData(0xf134, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF135 = IconData(0xf135, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF136 = IconData(0xf136, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF137 = IconData(0xf137, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF138 = IconData(0xf138, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF139 = IconData(0xf139, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF13A = IconData(0xf13a, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF13B = IconData(0xf13b, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF13C = IconData(0xf13c, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF13D = IconData(0xf13d, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF13E = IconData(0xf13e, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF13F = IconData(0xf13f, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF140 = IconData(0xf140, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF141 = IconData(0xf141, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF142 = IconData(0xf142, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF143 = IconData(0xf143, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF144 = IconData(0xf144, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF145 = IconData(0xf145, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF146 = IconData(0xf146, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF147 = IconData(0xf147, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF148 = IconData(0xf148, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF149 = IconData(0xf149, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF14A = IconData(0xf14a, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF14B = IconData(0xf14b, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF14C = IconData(0xf14c, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF14D = IconData(0xf14d, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF14E = IconData(0xf14e, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF14F = IconData(0xf14f, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF150 = IconData(0xf150, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF151 = IconData(0xf151, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF152 = IconData(0xf152, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF153 = IconData(0xf153, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF154 = IconData(0xf154, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF155 = IconData(0xf155, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF156 = IconData(0xf156, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF157 = IconData(0xf157, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF158 = IconData(0xf158, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF159 = IconData(0xf159, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF15A = IconData(0xf15a, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF15B = IconData(0xf15b, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF15C = IconData(0xf15c, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF15D = IconData(0xf15d, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF15E = IconData(0xf15e, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF15F = IconData(0xf15f, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF160 = IconData(0xf160, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF161 = IconData(0xf161, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF162 = IconData(0xf162, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF163 = IconData(0xf163, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF164 = IconData(0xf164, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF165 = IconData(0xf165, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF166 = IconData(0xf166, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF167 = IconData(0xf167, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF168 = IconData(0xf168, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF169 = IconData(0xf169, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF16A = IconData(0xf16a, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF16B = IconData(0xf16b, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF16C = IconData(0xf16c, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF16D = IconData(0xf16d, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF16E = IconData(0xf16e, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF16F = IconData(0xf16f, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF170 = IconData(0xf170, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF171 = IconData(0xf171, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF172 = IconData(0xf172, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF173 = IconData(0xf173, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF174 = IconData(0xf174, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF175 = IconData(0xf175, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF176 = IconData(0xf176, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF177 = IconData(0xf177, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF178 = IconData(0xf178, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF179 = IconData(0xf179, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF17A = IconData(0xf17a, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF17B = IconData(0xf17b, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF17C = IconData(0xf17c, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF17D = IconData(0xf17d, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF17E = IconData(0xf17e, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF17F = IconData(0xf17f, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF180 = IconData(0xf180, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF181 = IconData(0xf181, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF182 = IconData(0xf182, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF183 = IconData(0xf183, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF184 = IconData(0xf184, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF185 = IconData(0xf185, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF186 = IconData(0xf186, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF187 = IconData(0xf187, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF188 = IconData(0xf188, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF189 = IconData(0xf189, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF18A = IconData(0xf18a, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF18B = IconData(0xf18b, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF18C = IconData(0xf18c, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF18D = IconData(0xf18d, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF18E = IconData(0xf18e, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF18F = IconData(0xf18f, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF190 = IconData(0xf190, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF191 = IconData(0xf191, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF192 = IconData(0xf192, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF193 = IconData(0xf193, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF194 = IconData(0xf194, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF195 = IconData(0xf195, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF196 = IconData(0xf196, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF197 = IconData(0xf197, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF198 = IconData(0xf198, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF199 = IconData(0xf199, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF19A = IconData(0xf19a, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF19B = IconData(0xf19b, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF19C = IconData(0xf19c, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF19D = IconData(0xf19d, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF19E = IconData(0xf19e, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF19F = IconData(0xf19f, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF1A0 = IconData(0xf1a0, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF1A1 = IconData(0xf1a1, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF1A2 = IconData(0xf1a2, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF1A3 = IconData(0xf1a3, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF1A4 = IconData(0xf1a4, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF1A5 = IconData(0xf1a5, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF1A6 = IconData(0xf1a6, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF1A7 = IconData(0xf1a7, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF1A8 = IconData(0xf1a8, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF1A9 = IconData(0xf1a9, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF1AA = IconData(0xf1aa, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF1AB = IconData(0xf1ab, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF1AC = IconData(0xf1ac, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF1AD = IconData(0xf1ad, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF1AE = IconData(0xf1ae, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF1AF = IconData(0xf1af, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF1B0 = IconData(0xf1b0, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF1B1 = IconData(0xf1b1, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF1B2 = IconData(0xf1b2, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF1B3 = IconData(0xf1b3, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF1B4 = IconData(0xf1b4, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF1B5 = IconData(0xf1b5, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF1B6 = IconData(0xf1b6, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF1B7 = IconData(0xf1b7, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF1B8 = IconData(0xf1b8, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF1B9 = IconData(0xf1b9, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF1BA = IconData(0xf1ba, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF1BB = IconData(0xf1bb, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF1BC = IconData(0xf1bc, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF1BD = IconData(0xf1bd, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF1BE = IconData(0xf1be, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF1BF = IconData(0xf1bf, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF1C0 = IconData(0xf1c0, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF1C1 = IconData(0xf1c1, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF1C2 = IconData(0xf1c2, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF1C3 = IconData(0xf1c3, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF1C4 = IconData(0xf1c4, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF1C5 = IconData(0xf1c5, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF1C6 = IconData(0xf1c6, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF1C7 = IconData(0xf1c7, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF1C8 = IconData(0xf1c8, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF1C9 = IconData(0xf1c9, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF1CA = IconData(0xf1ca, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF1CB = IconData(0xf1cb, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF1CC = IconData(0xf1cc, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF1CD = IconData(0xf1cd, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF1CE = IconData(0xf1ce, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF1CF = IconData(0xf1cf, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF1D0 = IconData(0xf1d0, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF1D1 = IconData(0xf1d1, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF1D2 = IconData(0xf1d2, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF1D3 = IconData(0xf1d3, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF1D4 = IconData(0xf1d4, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF1D5 = IconData(0xf1d5, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF1D6 = IconData(0xf1d6, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF1D7 = IconData(0xf1d7, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF1D8 = IconData(0xf1d8, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF1D9 = IconData(0xf1d9, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF1DA = IconData(0xf1da, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF1DB = IconData(0xf1db, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF1DC = IconData(0xf1dc, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF1DD = IconData(0xf1dd, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF1DE = IconData(0xf1de, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF1DF = IconData(0xf1df, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF1E0 = IconData(0xf1e0, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF1E1 = IconData(0xf1e1, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF1E2 = IconData(0xf1e2, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF1E3 = IconData(0xf1e3, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF1E4 = IconData(0xf1e4, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF1E5 = IconData(0xf1e5, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF1E6 = IconData(0xf1e6, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF1E7 = IconData(0xf1e7, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF1E8 = IconData(0xf1e8, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF1E9 = IconData(0xf1e9, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF1EA = IconData(0xf1ea, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF1EB = IconData(0xf1eb, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF1EC = IconData(0xf1ec, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF1ED = IconData(0xf1ed, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF1EE = IconData(0xf1ee, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF1EF = IconData(0xf1ef, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF1F0 = IconData(0xf1f0, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF1F1 = IconData(0xf1f1, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF1F2 = IconData(0xf1f2, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF1F3 = IconData(0xf1f3, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF1F4 = IconData(0xf1f4, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF1F5 = IconData(0xf1f5, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF1F6 = IconData(0xf1f6, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF1F7 = IconData(0xf1f7, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF1F8 = IconData(0xf1f8, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF1F9 = IconData(0xf1f9, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF1FA = IconData(0xf1fa, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF1FB = IconData(0xf1fb, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF1FC = IconData(0xf1fc, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF1FD = IconData(0xf1fd, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF1FE = IconData(0xf1fe, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF1FF = IconData(0xf1ff, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF200 = IconData(0xf200, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF201 = IconData(0xf201, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF202 = IconData(0xf202, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF203 = IconData(0xf203, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF204 = IconData(0xf204, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF205 = IconData(0xf205, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF206 = IconData(0xf206, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF207 = IconData(0xf207, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF208 = IconData(0xf208, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF209 = IconData(0xf209, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF20A = IconData(0xf20a, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF20B = IconData(0xf20b, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF20C = IconData(0xf20c, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF20D = IconData(0xf20d, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF20E = IconData(0xf20e, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF20F = IconData(0xf20f, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF210 = IconData(0xf210, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF211 = IconData(0xf211, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF212 = IconData(0xf212, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF213 = IconData(0xf213, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF214 = IconData(0xf214, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF215 = IconData(0xf215, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF216 = IconData(0xf216, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF217 = IconData(0xf217, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF218 = IconData(0xf218, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF219 = IconData(0xf219, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF21A = IconData(0xf21a, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF21B = IconData(0xf21b, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF21C = IconData(0xf21c, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF21D = IconData(0xf21d, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF21E = IconData(0xf21e, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF21F = IconData(0xf21f, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF220 = IconData(0xf220, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF221 = IconData(0xf221, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF222 = IconData(0xf222, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF223 = IconData(0xf223, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF224 = IconData(0xf224, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF225 = IconData(0xf225, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF226 = IconData(0xf226, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF227 = IconData(0xf227, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF228 = IconData(0xf228, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF229 = IconData(0xf229, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF22A = IconData(0xf22a, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF22B = IconData(0xf22b, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF22C = IconData(0xf22c, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF22D = IconData(0xf22d, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF22E = IconData(0xf22e, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF22F = IconData(0xf22f, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF230 = IconData(0xf230, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF231 = IconData(0xf231, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF232 = IconData(0xf232, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF233 = IconData(0xf233, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF234 = IconData(0xf234, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF235 = IconData(0xf235, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF236 = IconData(0xf236, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF237 = IconData(0xf237, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF238 = IconData(0xf238, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF239 = IconData(0xf239, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF23A = IconData(0xf23a, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF23B = IconData(0xf23b, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF23C = IconData(0xf23c, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF23D = IconData(0xf23d, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF23E = IconData(0xf23e, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF23F = IconData(0xf23f, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF240 = IconData(0xf240, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF241 = IconData(0xf241, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF242 = IconData(0xf242, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF243 = IconData(0xf243, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData uniF244 = IconData(0xf244, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-}
