+++ lib/navigation/router.dart
@@ -3,6 +3,8 @@ import 'package:flutter/material.dart';
 import 'package:myapp/main.dart';
 import 'package:myapp/navigation/route_paths.dart';
 import 'package:myapp/screens/LoginSignUpPage.dart';
+import 'package:myapp/screens/homePage.dart';
+import 'package:myapp/screens/login_page.dart';
 import 'package:myapp/screens/profileSetUp.dart';
 
 class Router {
@@ -11,7 +13,7 @@ class Router {
       case RoutePaths.Home:
         return MaterialPageRoute(builder: (_) => HomePage());
       case RoutePaths.Login:
-        return MaterialPageRoute(builder: (_) => LoginSignupPage());
+        return MaterialPageRoute(builder: (_) => LoginPage());
       case RoutePaths.LoginProfile:
         return MaterialPageRoute(builder: (_) => ProfileSetUp());
     }
