+++ lib/providers/user_provider.dart
@@ -1,0 +1,21 @@
+
+import 'package:flutter/cupertino.dart';
+import 'package:myapp/services/firebase_authentication_service.dart';
+
+class UserProvider extends ChangeNotifier{
+  FirebaseAuthenticationService _firebaseAuthenticationService;
+
+  UserProvider(){
+    _firebaseAuthenticationService = FirebaseAuthenticationService();
+  }
+   Future<String> login(String email, String password) async{
+    return await _firebaseAuthenticationService.signIn(email, password);
+  }
+  Future<bool> isLoggedIn() async{
+    return await _firebaseAuthenticationService.getCurrentUser() != null;
+  }
+  logOut(){
+    _firebaseAuthenticationService.signOut();
+  }
+
+}
