+++ lib/services/firebase_authentication_service.dart
@@ -1,0 +1,70 @@
+import 'dart:async';
+import 'package:firebase_auth/firebase_auth.dart';
+
+abstract class BaseAuth {
+  Future<String> signIn(String email, String password);
+
+  Future<String> signUp(String email, String password);
+
+  Future<FirebaseUser> getCurrentUser();
+
+  Future<void> sendEmailVerification();
+
+  Future<void> signOut();
+
+  Future<bool> isEmailVerified();
+}
+class FirebaseAuthenticationService implements BaseAuth {
+  FirebaseAuth _firebaseAuth;
+
+  FirebaseAuthenticationService() {
+   _firebaseAuth = FirebaseAuth.instance;
+  }
+
+  Future<String> signIn(String email, String password) async {
+    FirebaseUser user;
+    AuthResult result;
+try {
+   result = await _firebaseAuth.signInWithEmailAndPassword(
+      email: email, password: password);
+}catch(Exception){
+  return null;
+}
+     user = result.user;
+
+    return user.uid;
+  }
+
+  Future<String> signUp(String email, String password) async {
+    FirebaseUser user;
+    try {
+      AuthResult result = await _firebaseAuth.createUserWithEmailAndPassword(
+          email: email, password: password);
+       user = result.user;
+    }catch(exception){
+      return null;
+    }
+    return user.uid;
+  }
+
+  Future<FirebaseUser> getCurrentUser() async {
+      FirebaseUser user = await _firebaseAuth.currentUser();
+      print(user.toString());
+      return user;
+
+  }
+
+  Future<void> signOut() async {
+    return _firebaseAuth.signOut();
+  }
+
+  Future<void> sendEmailVerification() async {
+    FirebaseUser user = await _firebaseAuth.currentUser();
+    user.sendEmailVerification();
+  }
+
+  Future<bool> isEmailVerified() async {
+    FirebaseUser user = await _firebaseAuth.currentUser();
+    return user.isEmailVerified;
+  }
+}
