+++ lib/tools/router_paths.dart
@@ -1,0 +1,6 @@
+class RoutePaths{
+  static const String Home = '/';
+  static const String Login = 'login';
+  static const String SignUp = "signup";
+
+}
