+++ pubspec.yaml
@@ -18,6 +18,8 @@ environment:
 
 dependencies:
   firebase_auth: ^0.16.1
+  carousel_slider: ^1.4.1
+  provider: ^4.3.2+3
   shared_preferences: ^0.5.12
   firebase_database: ^3.1.6
   animated_text_kit: ^2.2.0
@@ -65,6 +67,9 @@ flutter:
   # example:
 
   fonts:
+    - family: SFPro
+      fonts:
+        - asset: fonts/SF-Pro.ttf
     - family: Comfortaa
       fonts:
         - asset: fonts/Comfortaa-Light.ttf
